+++ .github/workflows/release.yml
@@ -31,7 +31,7 @@ jobs:
       MOVING_TAG: ${{ inputs.release_type == 'stable release' && 'latest' || 'next' }}
     steps:
       - name: Checkout
-        uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
+        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
         with:
           fetch-depth: 0
 
@@ -245,7 +245,7 @@ jobs:
       # Log in before the existence probes: anonymous `docker manifest inspect`
       # is rate limited, and a 429 is indistinguishable from "missing" below.
       - name: Login to Docker Hub
-        uses: docker/login-action@650006c6eb7dba73a995cc03b0b2d7f5ca915bee # v4.2.0
+        uses: docker/login-action@af1e73f918a031802d376d3c8bbc3fe56130a9b0 # v4.4.0
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
@@ -270,7 +270,7 @@ jobs:
 
       - name: Build and push Flow image
         if: steps.check-flow.outputs.exists == 'false'
-        uses: docker/build-push-action@f9f3042f7e2789586610d6e8b85c8f03e5195baf # v7.2.0
+        uses: docker/build-push-action@53b7df96c91f9c12dcc8a07bcb9ccacbed38856a # v7.3.0
         with:
           context: ./packages/cli
           file: ./packages/cli/Dockerfile
@@ -292,7 +292,7 @@ jobs:
 
       - name: Build and push CLI image
         if: steps.check-cli.outputs.exists == 'false'
-        uses: docker/build-push-action@f9f3042f7e2789586610d6e8b85c8f03e5195baf # v7.2.0
+        uses: docker/build-push-action@53b7df96c91f9c12dcc8a07bcb9ccacbed38856a # v7.3.0
         with:
           context: ./packages/cli
           file: ./packages/cli/Dockerfile.cli
