+++ package.json
@@ -61,7 +61,7 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.30.0",
-    "@walkeros/config": "4.1.0",
+    "@walkeros/config": "^4.3.0",
     "gray-matter": "^4.0.3",
     "husky": "^9.1.7",
     "lint-staged": "^17.0.7",
